scrape_fights included in Scraper class
