--- src/fight.py
@@ -1,7 +1,0 @@
-class Fight:
-    def __init__(self, white, blue, competition, date, round_cat):
-        self.white = white
-        self.blue = blue
-        self.competition = competition
-        self.date = date
-        self.round_cat = round_cat
